08_tools_web_search/pyproject.toml
[project]
name = "08-tools-web-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.5.5",
    "grpcio<1.72.1",
    "mypy>=1.16.0",
    "openai-agents>=0.0.17",
    "python-dotenv>=1.1.0",
    "selenium>=4.33.0",
    "webdriver-manager>=4.0.2",
]
